requirements.txt
langgraph>=0.2.50
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0,<3.0.0 